Extract discover components
